Prediction for puddle and spillable examines
