+++ Content.Server/Destructible/Thresholds/Behaviors/SpillBehavior.cs
@@ -1,6 +1,6 @@
 using Content.Server.Chemistry.Containers.EntitySystems;
-using Content.Server.Fluids.Components;
 using Content.Server.Fluids.EntitySystems;
+using Content.Shared.Fluids.Components;
 using JetBrains.Annotations;
 
 namespace Content.Server.Destructible.Thresholds.Behaviors
