+++ Content.Shared/Fluids/Components/EvaporationComponent.cs
@@ -1,19 +1,20 @@
-﻿using Content.Server.Fluids.EntitySystems;
-using Content.Shared.FixedPoint;
+﻿using Content.Shared.FixedPoint;
+using Robust.Shared.GameStates;
 using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom;
 
-namespace Content.Server.Fluids.Components;
+namespace Content.Shared.Fluids.Components;
 
 /// <summary>
 /// Added to puddles that contain water so it may evaporate over time.
 /// </summary>
-[RegisterComponent, Access(typeof(PuddleSystem))]
+[NetworkedComponent]
+[RegisterComponent, Access(typeof(SharedPuddleSystem))]
 public sealed partial class EvaporationComponent : Component
 {
     /// <summary>
     /// The next time we remove the EvaporationSystem reagent amount from this entity.
     /// </summary>
-    [ViewVariables(VVAccess.ReadWrite), DataField("nextTick", customTypeSerializer:typeof(TimeOffsetSerializer))]
+    [ViewVariables(VVAccess.ReadWrite), DataField("nextTick", customTypeSerializer: typeof(TimeOffsetSerializer))]
     public TimeSpan NextTick = TimeSpan.Zero;
 
     /// <summary>
