+++ Content.Shared/Fluids/Components/PuddleComponent.cs
@@ -11,10 +11,10 @@ namespace Content.Shared.Fluids.Components
     [RegisterComponent, NetworkedComponent, Access(typeof(SharedPuddleSystem))]
     public sealed partial class PuddleComponent : Component
     {
-        [DataField("spillSound")]
+        [DataField]
         public SoundSpecifier SpillSound = new SoundPathSpecifier("/Audio/Effects/Fluids/splat.ogg");
 
-        [DataField("overflowVolume")]
+        [DataField]
         public FixedPoint2 OverflowVolume = FixedPoint2.New(20);
 
         [DataField("solution")] public string SolutionName = "puddle";
