+++ Content.Shared/Fluids/SharedPuddleSystem.Evaporation.cs
@@ -1,0 +1,17 @@
+using Content.Shared.Chemistry.Components;
+using Content.Shared.Chemistry.Reagent;
+
+namespace Content.Shared.Fluids;
+
+public abstract partial class SharedPuddleSystem
+{
+    [ValidatePrototypeId<ReagentPrototype>]
+    private const string Water = "Water";
+
+    public static readonly string[] EvaporationReagents = [Water];
+
+    public bool CanFullyEvaporate(Solution solution)
+    {
+        return solution.GetTotalPrototypeQuantity(EvaporationReagents) == solution.Volume;
+    }
+}
