+++ Resources/Locale/en-US/fluids/components/puddle-component.ftl
@@ -1,4 +1,4 @@
-puddle-component-examine-is-slipper-text = It looks [color=#169C9C]slippery[/color].
+puddle-component-examine-is-slippery-text = It looks [color=#169C9C]slippery[/color].
 puddle-component-examine-evaporating = It is [color=#5E7C16]evaporating[/color].
 puddle-component-examine-evaporating-partial = It is [color=#FED83D]partially evaporating[/color].
 puddle-component-examine-evaporating-no = It is [color=#B02E26]not evaporating[/color].
